refactor: semi working refactor of editor page
