+++ apps/desktop/package.json
@@ -34,7 +34,7 @@
     "export-to-csv": "^1.4.0",
     "fuse.js": "^7.1.0",
     "hotkeys-js": "^3.13.7",
-    "lucide-react": "^0.525.0",
+    "lucide-react": "^1.14.0",
     "mini-svg-data-uri": "^1.4.4",
     "nanoid": "^5.1.5",
     "react": "catalog:react",
