+++ apps/desktop/src/components/data-table/data-table.tsx
@@ -5,6 +5,8 @@ import {
 } from "@tanstack/react-table"
 import React, { useMemo } from "react"
 
+import { getCommonPinningStyles } from "@/lib/data-table"
+import { cn } from "@tablex/lib/utils"
 import {
   Table,
   TableBody,
@@ -13,9 +15,6 @@ import {
   TableHeader,
   TableRow
 } from "@tablex/ui/components/table"
-import { getCommonPinningStyles } from "@/lib/data-table"
-import { cn } from "@tablex/lib/utils"
-import { ScrollArea, ScrollBar } from "@tablex/ui/components/scroll-area"
 
 interface DataTableProps<TData> extends React.ComponentProps<"div"> {
   table: TanstackTable<TData>
