--- apps/desktop/src/routes/connection.$connId/_table-layout/table-view/empty/route.tsx
@@ -1,18 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router"
-
-export const Route = createFileRoute(
-  "/connection/$connId/_table-layout/table-view/empty"
-)({
-  component: DashboardPage
-})
-
-function DashboardPage() {
-  return (
-    <div className="flex h-full w-full flex-col items-center justify-center gap-y-3 wrap-break-word text-center text-2xl font-bold text-gray-500 opacity-50">
-      <img src={"/icons/cube.svg"} alt="cube" width={100} height={100} />
-      Choose a table
-      <br />
-      to get started
-    </div>
-  )
-}
